refactor: 重構 socket 事件管理
